Add the GetModelLabels function
Add the GetModelLabels function and corresponding LabelsReady
event. When used, these provide a list of the known classification
labels in the model.

Updated version component version to 20210224. Note: This is not
original work. This function/event was found in a compiled binary of (as
of today) unknown provenance. The changes were reverse engineered and
applied here.